Owners

+++ configuration/views.py
@@ -1,15 +1,25 @@
-from django.http import JsonResponse, Http404
-from django.shortcuts import render, redirect
-from django.views.generic import CreateView, DeleteView, ListView, DetailView, UpdateView
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.sites.shortcuts import get_current_site
+from django.core.mail import EmailMessage, send_mail
+from django.db.models import ProtectedError
+from django.http import Http404
+from django.template.loader import render_to_string
+from django.utils.encoding import force_bytes, force_str
+from django.utils.http import urlsafe_base64_encode, urlsafe_base64_decode
 
-from .models import *
+import random
+
+from administrator_panel.mixins import *
 from .forms import *
+from .tokens import *
 
 
-class MeasurementUnitListView(CreateView):
+class MeasurementUnitListView(PermissionCreateView):
     model = MeasurementUnit
     form_class = MeasurementUnitForm
     template_name = 'configuration/measurement_unit-list.html'
+    string_permission = 'service_access'
 
     def get_context_data(self, **kwargs):
         measurement_unit_queryset = MeasurementUnit.objects.all().order_by('id')
@@ -34,10 +44,11 @@ class MeasurementUnitListView(CreateView):
         return self.render_to_response(context)
 
 
-class MeasurementUnitDeleteView(DeleteView):
+class MeasurementUnitDeleteView(PermissionDeleteView):
     model = MeasurementUnit
     template_name = 'configuration/measurement_unit-list.html'
     pk_url_kwarg = 'pk'
+    string_permission = 'service_access'
 
     def delete(self, request, *args, **kwargs):
         try:
@@ -51,10 +62,11 @@ class MeasurementUnitDeleteView(DeleteView):
         return JsonResponse({'answer': 'success'})
 
 
-class ServiceCreateListView(CreateView):
+class ServiceCreateListView(PermissionCreateView):
     model = Service
     form_class = ServiceForm
     template_name = 'configuration/service-list.html'
+    string_permission = 'service_access'
 
     def get_context_data(self, **kwargs):
         context = super(ServiceCreateListView, self).get_context_data(**kwargs)
@@ -80,10 +92,11 @@ class ServiceCreateListView(CreateView):
         return self.render_to_response(context)
 
 
-class ServiceDeleteView(DeleteView):
+class ServiceDeleteView(PermissionDeleteView):
     model = Service
     template_name = 'configuration/service-list.html'
     pk_url_kwarg = 'pk'
+    string_permission = 'service_access'
 
     def delete(self, request, *args, **kwargs):
         try:
@@ -95,15 +108,16 @@ class ServiceDeleteView(DeleteView):
             if not related_measurement_unit.service_set.all().exists():
                 related_measurement_unit.is_used = False
                 related_measurement_unit.save()
-        except Service.DoesNotExist:
+        except (Service.DoesNotExist, ProtectedError):
             return JsonResponse({'answer': 'failed'})
         return JsonResponse({'answer': 'success'})
 
 
-class TariffCreateView(CreateView):
+class TariffCreateView(PermissionCreateView):
     model = Tariff
     template_name = 'configuration/tariff-create-update.html'
     form_class = TariffForm
+    string_permission = 'tariff_access'
 
     def get_object(self, queryset=None):
         try:
@@ -148,15 +162,17 @@ class TariffCreateView(CreateView):
         return redirect('tariffs')
 
 
-class TariffListView(ListView):
+class TariffListView(PermissionListView):
     queryset = Tariff.objects.all().order_by('updated_at')
     template_name = 'configuration/tariff-list.html'
+    string_permission = 'tariff_access'
 
 
-class TariffDetailView(DetailView):
+class TariffDetailView(PermissionDetailView):
     model = Tariff
     template_name = 'configuration/tariff-detail.html'
     pk_url_kwarg = 'pk'
+    string_permission = 'tariff_access'
 
     def get_object(self, queryset=None):
         return Tariff.objects.prefetch_related('tariffservice_set',
@@ -164,11 +180,12 @@ class TariffDetailView(DetailView):
                                                'tariffservice_set__service__measurement_unit').get(pk=self.kwargs.get('pk'))
 
 
-class TariffUpdateView(UpdateView):
+class TariffUpdateView(PermissionUpdateView):
     model = Tariff
     template_name = 'configuration/tariff-create-update.html'
     pk_url_kwarg = 'pk'
     form_class = TariffForm
+    string_permission = 'tariff_access'
 
     def get_object(self, queryset=None):
         try:
@@ -215,6 +232,8 @@ class TariffUpdateView(UpdateView):
 
 
 def tariff_delete(request, pk):
+    if request.user.is_anonymous or not request.user.role.tariff_access:
+        return JsonResponse({'answer': 'Ви не маєте доступу до тарифів'})
     try:
         tariff_to_delete = Tariff.objects.get(pk=pk)
         tariff_to_delete.delete()
@@ -224,6 +243,8 @@ def tariff_delete(request, pk):
 
 
 def tariff_service_delete(request, pk_tariff_service_to_delete):
+    if request.user.is_anonymous or not request.user.role.tariff_access:
+        return JsonResponse({'answer': 'Ви не маєте доступу до тарифів'})
     try:
         tariff_service_to_delete = TariffService.objects.get(pk=pk_tariff_service_to_delete)
         tariff_service_to_delete.delete()
@@ -232,9 +253,10 @@ def tariff_service_delete(request, pk_tariff_service_to_delete):
     return JsonResponse({'answer': 'success'})
 
 
-class RolesListUpdate(ListView):
+class RolesListUpdate(PermissionListView):
     model = Role
     template_name = 'configuration/role-list.html'
+    string_permission = 'role_access'
 
     def get_context_data(self, *, object_list=None, **kwargs):
         context = super(RolesListUpdate, self).get_context_data()
@@ -255,10 +277,11 @@ class RolesListUpdate(ListView):
         return self.render_to_response(context)
 
 
-class UserCreateView(CreateView):
+class UserCreateView(PermissionCreateView):
     model = User
     template_name = 'configuration/user-create-update.html'
     form_class = UserForm
+    string_permission = 'user_access'
 
     def get_context_data(self, **kwargs):
         context = super(UserCreateView, self).get_context_data(**kwargs)
@@ -285,10 +308,11 @@ class UserCreateView(CreateView):
                                  role=form.cleaned_data.get('role'))
 
 
-class UserDetailView(DetailView):
+class UserDetailView(PermissionDetailView):
     model = User
     template_name = 'configuration/user-detail.html'
     pk_url_kwarg = 'pk'
+    string_permission = 'user_access'
 
     def get_object(self, queryset=None):
         try:
@@ -297,11 +321,12 @@ class UserDetailView(DetailView):
             raise Http404()
 
 
-class UserUpdateView(UpdateView):
+class UserUpdateView(PermissionUpdateView):
     model = User
     template_name = 'configuration/user-create-update.html'
     form_class = UserForm
     pk_url_kwarg = 'pk'
+    string_permission = 'user_access'
 
     def get_object(self, queryset=None):
         try:
@@ -339,15 +364,18 @@ class UserUpdateView(UpdateView):
         return redirect('users')
 
 
-class UserListView(ListView):
+class UserListView(PermissionListView):
     model = User
     template_name = 'configuration/user-list.html'
+    string_permission = 'user_access'
 
     def get_queryset(self):
         return User.objects.select_related('role').all().order_by('pk')
 
 
 def delete_user(request, pk):
+    if request.user.is_anonymous or not request.user.role.user_access:
+        return JsonResponse({'answer': 'Ви не маєте доступу до користувачів'})
     try:
         user = User.objects.get(pk=pk)
         user.delete()
@@ -356,10 +384,11 @@ def delete_user(request, pk):
     return JsonResponse({'answer': 'success'})
 
 
-class PaymentRequisitesCreateView(CreateView):
+class PaymentRequisitesCreateView(PermissionCreateView):
     model = PaymentRequisite
     template_name = 'configuration/payment-requisite.html'
     form_class = PaymentRequisitesForm
+    string_permission = 'payment_requisite_access'
 
     def get_form(self, form_class=None):
         return PaymentRequisitesForm(instance=PaymentRequisite.objects.first())
@@ -373,10 +402,11 @@ class PaymentRequisitesCreateView(CreateView):
         return self.render_to_response(self.get_context_data())
 
 
-class ArticlePaymentCreateView(CreateView):
+class ArticlePaymentCreateView(PermissionCreateView):
     model = ArticlePayment
     template_name = 'configuration/article-payment-create-update.html'
     form_class = ArticlePaymentForm
+    string_permission = 'payment_requisite_access'
 
     def get_form(self, form_class=None):
         return ArticlePaymentForm()
@@ -392,17 +422,19 @@ class ArticlePaymentCreateView(CreateView):
         return self.render_to_response(context)
 
 
-class ArticlePaymentListView(ListView):
+class ArticlePaymentListView(PermissionListView):
     model = ArticlePayment
     template_name = 'configuration/article-payment-list.html'
     queryset = ArticlePayment.objects.all().order_by('pk')
+    string_permission = 'payment_requisite_access'
 
 
-class ArticlePaymentUpdateView(UpdateView):
+class ArticlePaymentUpdateView(PermissionUpdateView):
     model = ArticlePayment
     template_name = 'configuration/article-payment-create-update.html'
     form_class = ArticlePaymentForm
     pk_url_kwarg = 'pk'
+    string_permission = 'payment_requisite_access'
 
     def get_object(self, queryset=None):
         try:
@@ -422,9 +454,179 @@ class ArticlePaymentUpdateView(UpdateView):
 
 
 def delete_article(request, pk):
+    if request.user.is_anonymous or not request.user.role.payment_requisite_access:
+        return JsonResponse({'answer': 'Ви не маєте доступу до користувачів'})
     try:
         article_to_delete = ArticlePayment.objects.get(pk=pk)
         article_to_delete.delete()
-    except ArticlePayment.DoesNotExist:
+    except (ArticlePayment.DoesNotExist, ProtectedError):
         return JsonResponse({'answer': 'failed'})
     return JsonResponse({'answer': 'success'})
+
+
+class UserLoginView(CreateView):
+    model = User
+    template_name = 'configuration/user-login.html'
+    form_class = UserLoginForm
+
+    def get(self, request, *args, **kwargs):
+        if request.user.is_authenticated:
+            if request.user.role.role == 'owner':
+                return redirect('owner-receipts')
+            return redirect('statistics')
+        self.object = None
+        context = self.get_context_data()
+        return self.render_to_response(context)
+
+    def post(self, request, *args, **kwargs):
+        email = request.POST.get('email')
+        password = request.POST.get('password')
+        if email and password:
+            try:
+                user = User.objects.get(email=email)
+            except User.DoesNotExist:
+                return self.errors_occurred()
+
+            if not user.check_password(password):
+                user = None
+
+            if user and user.role.role == 'owner' and user.status != 'disconnected':
+                if not user.is_active:
+                    messages.error(request, 'Ваш обліковий запис не підтверджений! Перегляньте ваші листи на пошті та підтвердіть його.')
+                    return redirect('user-login')
+                login(request, user)
+                if not request.POST.get('remember_me'):
+                    self.request.session.set_expiry(0)
+                return redirect('owner-receipts')
+        self.object = None
+        context = self.get_context_data()
+        context['error'] = 'Неправильно введені дані'
+        return self.render_to_response(context=context)
+
+    def errors_occurred(self):
+        self.object = None
+        context = self.get_context_data()
+        context['error'] = 'Неправильно введені дані'
+        return self.render_to_response(context=context)
+
+
+class ManagementLoginView(CreateView):
+    model = User
+    template_name = 'configuration/user-staff-login.html'
+    form_class = UserLoginForm
+
+    def get(self, request, *args, **kwargs):
+        if self.request.user.is_authenticated:
+            if request.user.role.role == 'owner':
+                return redirect('owner-receipts')
+            return redirect('statistics')
+        self.object = None
+        context = self.get_context_data()
+        return self.render_to_response(context)
+
+    def post(self, request, *args, **kwargs):
+        email = request.POST.get('email')
+        password = request.POST.get('password')
+        if email and password:
+            user = authenticate(username=email,
+                                password=password)
+            if user and user.role.role != 'owner' and user.status != 'disconnected':
+                login(request, user)
+                if not request.POST.get('remember_me'):
+                    self.request.session.set_expiry(0)
+                return redirect('statistics')
+        self.object = None
+        context = self.get_context_data()
+        context['error'] = 'Неправильно введені дані'
+        return self.render_to_response(context=context)
+
+
+class UserLogoutView(View):
+    def get(self, request, *args, **kwargs):
+        if request.user.is_authenticated:
+            logout(request)
+        return redirect('user-login')
+
+
+def send_activation_mail(request, user, to_mail):
+    message_subject = 'Підтвердіть ваш обліковий запис.'
+    message_with_template = render_to_string("email-activation-template.html", {
+        'user': user.email,
+        'domain': get_current_site(request).domain,
+        'uid': urlsafe_base64_encode(force_bytes(user.pk)),
+        'token': account_activation_token.make_token(user),
+        "protocol": 'https' if request.is_secure() else 'http'
+    })
+
+    email = EmailMessage(message_subject, message_with_template, to=[to_mail])
+    email.content_subtype = 'html'
+    if email.send(fail_silently=True):
+        messages.success(request, 'Користувач зареєстрований. Підтвердіть із вказаної пошти ваш обліковий запис для закінчення реєстрації')
+    else:
+        messages.error(request, 'Не вдалося відправити підтвердження на пошту. Перевірте правильність введення email')
+
+
+class UserActivation(View):
+
+    def get(self, request, *args, **kwargs):
+        try:
+            uid = force_str(urlsafe_base64_decode(self.kwargs.get('uidb64')))
+            user = User.objects.get(pk=uid)
+        except:
+            user = None
+
+        if user is not None and account_activation_token.check_token(user, self.kwargs.get('token')):
+            user.is_active = True
+            user.save()
+            messages.success(self.request, 'Ваш облікований запис підтверджено. Ввійдіть у свій обліковий запис знову.')
+            return redirect('user-login')
+        else:
+            messages.error(self.request, 'Підтвердження облікового запису не відбулося.')
+        return redirect('user-login')
+
+
+class UserRegistrationView(CreateView):
+    model = User
+    template_name = 'configuration/user-registration.html'
+    form_class = UserSelfRegistrationForm
+
+    @staticmethod
+    def rand_x_digit_num(x: int, leading_zeroes=True):
+        """Return an X digit number, leading_zeroes returns a string, otherwise int"""
+        if not leading_zeroes:
+            return str(random.randint(10 ** (x - 1), 10 ** x - 1))
+        else:
+            if x > 6000:
+                return ''.join([str(random.randint(0, 9)) for i in range(x)])
+            else:
+                return '{0:0{x}d}'.format(random.randint(0, 10 ** x - 1), x=x)
+
+    def get(self, request, *args, **kwargs):
+        if self.request.user.is_authenticated:
+            if self.request.user.role.role == 'owner':
+                return redirect('owner-receipts')
+            return redirect('statistics')
+        return super().get(request, *args, **kwargs)
+
+    def post(self, request, *args, **kwargs):
+        form = self.get_form_class()(request.POST)
+        if form.is_valid():
+            return self.form_valid(form)
+        return self.form_invalid(form)
+
+    def form_valid(self, form):
+        user = User.objects.create_user(owner_id=UserRegistrationView.rand_x_digit_num(8),
+                                        email=form.cleaned_data.get('email'),
+                                        name=form.cleaned_data.get('name'),
+                                        surname=form.cleaned_data.get('surname'),
+                                        password=form.cleaned_data.get('password'),
+                                        is_active=False,
+                                        role=Role.objects.get(role='owner'),)
+        send_activation_mail(self.request, user, user.email)
+        return redirect('user-registration')
+
+    def form_invalid(self, form):
+        messages.error(self.request, 'Щось пішло не так. Перевірте правильність вводу даних')
+        self.object = None
+        context = self.get_context_data()
+        return self.render_to_response(context)
